simplify migration test
