+++ qa/test-db-instance-migration/test-migration/pom.xml
@@ -76,11 +76,6 @@
       <scope>test</scope>
     </dependency>
 
-    <dependency>
-      <groupId>org.hibernate.javax.persistence</groupId>
-      <artifactId>hibernate-jpa-2.0-api</artifactId>
-    </dependency>
-
   </dependencies>
 
   <build>
