+++ qa/test-db-instance-migration/test-migration/src/main/resources/camunda.cfg.xml
@@ -15,7 +15,7 @@
     <property name="databaseSchemaUpdate" value="true" />
 
     <property name="javaSerializationFormatEnabled" value="true" />
-    
+
     <!-- job executor configurations -->
     <property name="jobExecutorActivate" value="false" />
