--- qa/test-db-instance-migration/test-migration/src/test/java/org/camunda/bpm/qa/upgrade/scenarios7190/variables/EntityManagerSession.java
@@ -1,34 +1,0 @@
-/*
- * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH
- * under one or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information regarding copyright
- * ownership. Camunda licenses this file to you under the Apache License,
- * Version 2.0; you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.camunda.bpm.qa.upgrade.scenarios7190.variables;
-
-import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
-import org.camunda.bpm.engine.ProcessEngineException;
-import org.camunda.bpm.engine.impl.interceptor.Session;
-
-/**
- * @author Frederik Heremans
- */
-public interface EntityManagerSession extends Session {
-  /**
-   * Get an {@link EntityManager} instance associated with this session.
-   * @throws ProcessEngineException when no {@link EntityManagerFactory} instance
-   * is configured for the process engine.
-   */
-  EntityManager getEntityManager();
-}
