+++ qa/test-db-instance-migration/test-migration/src/test/java/org/camunda/bpm/qa/upgrade/scenarios7190/variables/JPAVariablesSerializerPlugin.java
@@ -16,8 +16,6 @@
  */
 package org.camunda.bpm.qa.upgrade.scenarios7190.variables;
 
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
 import org.camunda.bpm.engine.impl.cfg.AbstractProcessEnginePlugin;
 import org.camunda.bpm.engine.impl.cfg.ProcessEngineConfigurationImpl;
 import org.camunda.bpm.engine.impl.variable.serializer.VariableSerializers;
@@ -27,28 +25,12 @@ public class JPAVariablesSerializerPlugin extends AbstractProcessEnginePlugin {
 
   @Override
   public void postInit(ProcessEngineConfigurationImpl processEngineConfiguration) {
-    EntityManagerFactory jpaEntityManagerFactory = Persistence.createEntityManagerFactory("activiti-jpa-pu");
-    processEngineConfiguration.getSessionFactories().put(EntityManagerSession.class,
-        new EntityManagerSessionFactory(jpaEntityManagerFactory, true, true));
-
     VariableSerializers variableSerializers = processEngineConfiguration.getVariableSerializers();
     int index = variableSerializers.getSerializerIndexByName(ValueType.BYTES.getName());
     if (index > -1) {
-      variableSerializers.addSerializer(new JPAVariableSerializer(), index);
+      variableSerializers.addSerializer(new JPATestVariableSerializer(), index);
     } else {
-      variableSerializers.addSerializer(new JPAVariableSerializer());
+      variableSerializers.addSerializer(new JPATestVariableSerializer());
     }
-
-//    EntityManager manager = jpaEntityManagerFactory.createEntityManager();
-//    manager.getTransaction().begin();
-//
-//    FieldAccessJPAEntity entityToQuery = new FieldAccessJPAEntity();
-//    entityToQuery.setId(1L);
-//    entityToQuery.setMyValue("value1");
-//    manager.persist(entityToQuery);
-//
-//    manager.flush();
-//    manager.getTransaction().commit();
-//    manager.close();
   }
 }
