--- qa/test-db-instance-migration/test-migration/src/test/resources/META-INF/persistence.xml
@@ -1,28 +1,0 @@
-<persistence xmlns="http://java.sun.com/xml/ns/persistence"
-	 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-     xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_1_0.xsd"
-     version="1.0"
->
-  <persistence-unit name="activiti-jpa-pu">
-    <provider>org.apache.openjpa.persistence.PersistenceProviderImpl</provider>
-    
-    <class>org.camunda.bpm.qa.upgrade.variables.FieldAccessJPAEntity</class>
-<!--    <class>org.camunda.bpm.qa.upgrade.variables.PropertyAccessJPAEntity</class>-->
-<!--    <class>org.camunda.bpm.qa.upgrade.variables.SubclassFieldAccessJPAEntity</class>-->
-<!--    <class>org.camunda.bpm.qa.upgrade.variables.SubclassPropertyAccessJPAEntity</class>-->
-<!--     Mapped superclasses -->
-<!--    <class>org.camunda.bpm.qa.upgrade.variables.MappedSuperClassFieldAccessJPAEntity</class>-->
-<!--    <class>org.camunda.bpm.qa.upgrade.variables.MappedSuperClassPropertyAccessJPAEntity</class>-->
-
-    <exclude-unlisted-classes>true</exclude-unlisted-classes>
-    <properties>
-      <property name="openjpa.ConnectionURL" value="jdbc:h2:./camunda-h2-default/process-engine;TRACE_LEVEL_FILE=0;DB_CLOSE_ON_EXIT=FALSE"/>
-      <property name="openjpa.ConnectionDriverName" value="org.h2.Driver"/>
-      <property name="openjpa.ConnectionUserName" value="sa"/>
-      <property name="openjpa.ConnectionPassword" value=""/>
-      <property name="openjpa.Log" value="DefaultLevel=WARN, Tool=INFO"/>
-      <property name="openjpa.jdbc.SynchronizeMappings" value="buildSchema(SchemaAction='add,deleteTableContents')"/>
-      <property name="openjpa.RuntimeUnenhancedClasses" value="supported" />
-    </properties>
-  </persistence-unit>
-</persistence>
